♻️ refactor : 컴포넌트 폴더 이동 및 개선 - _components 제거 - 불필요한 컴포넌트 제거 - layout파일에서 데이터를 모두 받도록 개선
