+++ src/app/(public)/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import Feature from '@/components/landing/Feature';
+import Hero from '@/components/landing/Hero';
+import Setting from '@/components/landing/Setting';
+
+export default function page() {
+  return (
+    <>
+      <Hero />
+      <Feature />
+      <Setting />
+    </>
+  );
+}
