--- src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <main>홈페이지입니다.</main>;
-}
