+++ src/components/landing/Setting.tsx
