+++ src/components/landing/SmoothScroll.tsx
