+++ src/components/landing/layout/type.tsx
